+++ CHANGELOG.md
@@ -33,6 +33,7 @@ AdminNeo 5.8.0
 - Edit: Fix editing rows with binary columns in tables without a unique key
 - Editor: Fix JavaScript error on the select page caused by the import form (regression from 5.5.0)
 - Database schema: Fix overlapping table boxes
+- Database schema: Allow dragging a table box also by its name
 - MySQL: Fix displaying backward relation links
 - PostgreSQL: Fix editing a record with a GENERATED ALWAYS AS IDENTITY column (fix #205, regression from 5.5.0)
 - PostgreSQL, MS SQL: Highlight primary keys in the database schema (regression from 5.1.0)